TESTES(tests de integracao para empresa adc

+++ src/tests/jest.setup.ts
@@ -1,11 +1,22 @@
 import mongoose from "mongoose";
 import dotenv from "dotenv";
+
 dotenv.config({ path: ".env.test" });
 
-afterEach(async () => {
+beforeAll(async () => {
+  if (mongoose.connection.readyState === 0) {
+    await mongoose.connect(process.env.MONGO_URI!);
+  }
+});
+
+beforeEach(async () => {
+  if (mongoose.connection.readyState !== 1) return;
+
   const db = mongoose.connection.db;
   if (!db) return;
+
   const collections = await db.collections();
+
   for (const collection of collections) {
     await collection.deleteMany({});
   }
